move vivante eula from weston to wayland-base-vivante
Make it mandatory for the user to accept Vivante EULA while building
wayland-base-vivante by specifying --build-arg ACCEPT_FSL_EULA=1
If Vivante EULA is not accepted at build time, wayland-base-vivante
fails to build. Add ACCEPT_FSL_EULA: 1 to docker build variables for
*-vivante images.

At runtime, the Vivante EULA script is invoked from /etc/profile so this
applies only to login shells. Failure to accept Vivante EULA will exit
the running continer.

The EULA that was previously bundled had different spacing than the one
included in the imx-gpu-viv binary installer. This one is an exact copy
of the EULA provided in the installer.

Related-to: #55039

Signed-off-by: Abdur Rehman <abdur.rehman@toradex.com>