base: create home directory for torizon user
We first opted to not create a home directory to avoid that application
store state in the container. However, some application really rely on
it, e.g. Chromium to create default profiles. We still can bind mount
it outside the container/use a Docker volume if we want to preserve the
state.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>