Detect SoC to descide if GPU should be used
Check SoC to descide whether the GPU should be used.

Note: There is no easy/generic way to detect a GPU. We could detect a
DRM supported GPU by checking if render nodes are available, but that
does not work for Vivante based devices. So the safest way is to just
assume that the device has a GPU and and maintain a list of SoC for
which we want to use the Pixman software renderer.

Related-to: TOR-192
Signed-off-by: Stefan Agner <stefan.agner@toradex.com>