weston: add touch2pointer option
Add an option to start touch2pointer with a path to the touch device
which you want to use as pointer input. Make sure you set
SUBSYSTEM=="input", KERNEL=="event0", ENV{LIBINPUT_IGNORE_DEVICE}="1"
on the touch device you want to use. So that weston doesn't use the
pointer and the touch device as input.

Related-to: TOR-1378
Signed-off-by: Stefan Eichenberger <stefan.eichenberger@toradex.com>