Add Toradex snapshot CI variables
Expose a USE_TORADEX_SNAPSHOT variable to CI. This variable has a
default value of 1. When this is set to 1 the snapshot defined by
TORADEX_SNAPSHOT will be used, otherwise the main Toradex feed will be
used.

Also expose the TORADEX_SNAPSHOT variable to CI. This variable defaults
to the latest Toradex feed snapshot. If specified, the particular
snapshot will be used otherwise the default snapshot will be used.

Use cases:
* Containers are built with the latest Toradex snapshot by default.
* To build with a particular Toradex feed snapshot, set TORADEX_SNAPSHOT
  to that snapshot
* To build with main Toradex feed, set USE_TORADEX_SNAPSHOT to 0.

Related-to: TOR-1656

Signed-off-by: Abdur Rehman <abdur.rehman@toradex.com>