qt5-wayland: change kms-script.sh to autodetect card
Now on TorizonCore 6 the card driver symlink has changed from 'card0' to
'card1' for Verdin iMX8M Mini, but it still remains 'card0' for
TorizonCore 5.

Because of this our QT5 test cases were failing, so I changed
kms-script.sh function 'set_dri_device' to automatically check which
card is the correct link, instead of maintaining a hardcoded list with
the link for every device.

Related-to: TOR-2798

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>