weston: fix entry.sh error when ENABLE_RDP is used
set -e was introduced in 6c521fae9 which causes the script to "Exit
immediately if a command exits with a non-zero status".

When ENABLE_RDP is used, the following error is observed and the script
exits:
    tee: /var/volatile/weston.log: No such file or directory

This happens because /var/volatile is not present when tee starts
execution. To fix this we simply move the creation of /var/volatile
before tee is supposed to execute.

Related-to: TOR-2976

Signed-off-by: Abdur Rehman <abdur.rehman@toradex.com>