base: add gpg key in keyring instead of trusted by apt
This commit changes how `apt` uses the Toradex key to authenticate
repositories. Adding to keyring and referencing in the `Signed-By`
property ensures that the Toradex key is only used for the Toradex
feeds, instead of for all feeds, which is the case if the key is
added to the trusted ones in `trusted.gpg.d`.

Related-to: TOR-3002

Signed-off-by: Leonardo Held <leonardo.held@toradex.com>